super-arc/server/requirements.txt
flask
Flask-CORS
langchain
mysql-connector-python 
langchain-core
langchain-community
python-dotenv